python-fastapi/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.14
    # via -r requirements.in
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
attrs==25.3.0
    # via aiohttp
azure-core==1.32.0
    # via
    #   azure-cosmos
    #   azure-identity
azure-cosmos==4.9.0
    # via -r requirements.in
azure-identity==1.21.0
    # via -r requirements.in
black==25.1.0
    # via -r requirements.in
certifi==2025.1.31
    # via
    #   -r requirements.in
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   black
    #   uvicorn
coverage[toml]==7.7.1
    # via pytest-cov
cryptography==44.0.2
    # via
    #   azure-identity
    #   msal
    #   pyjwt
dnspython==2.7.0
    # via
    #   -r requirements.in
    #   pymongo
docopt==0.6.2
    # via -r requirements.in
duckdb==1.2.1
    # via -r requirements.in
faker==37.1.0
    # via -r requirements.in
fastapi==0.115.12
    # via -r requirements.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via
    #   -r requirements.in
    #   httpcore
    #   hypercorn
    #   uvicorn
    #   wsproto
h2==4.2.0
    # via hypercorn
hpack==4.1.0
    # via h2
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via -r requirements.in
hypercorn==0.17.3
    # via -r requirements.in
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via -r requirements.in
markupsafe==3.0.2
    # via jinja2
msal==1.32.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.3.1
    # via azure-identity
multidict==6.2.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via black
packaging==24.2
    # via
    #   black
    #   pytest
pathspec==0.12.1
    # via black
platformdirs==4.3.7
    # via black
pluggy==1.5.0
    # via pytest
priority==2.0.0
    # via hypercorn
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
psutil==7.0.0
    # via -r requirements.in
pycparser==2.22
    # via cffi
pydantic==2.11.0
    # via fastapi
pydantic-core==2.33.0
    # via pydantic
pyjwt[crypto]==2.10.1
    # via
    #   msal
    #   pyjwt
pymongo==4.11.3
    # via -r requirements.in
pytest==8.3.5
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-cov
pytest-asyncio==0.26.0
    # via -r requirements.in
pytest-cov==6.0.0
    # via -r requirements.in
python-dotenv==1.1.0
    # via -r requirements.in
pytz==2025.2
    # via -r requirements.in
redis==5.2.1
    # via -r requirements.in
requests==2.32.3
    # via
    #   azure-core
    #   msal
six==1.17.0
    # via
    #   -r requirements.in
    #   azure-core
sniffio==1.3.1
    # via anyio
starlette==0.46.1
    # via fastapi
typing-extensions==4.13.0
    # via
    #   anyio
    #   azure-core
    #   azure-cosmos
    #   azure-identity
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
tzdata==2025.2
    # via faker
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via -r requirements.in
wsproto==1.2.0
    # via hypercorn
yarl==1.18.3
    # via aiohttp
